+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-17 17:52:53
+
+### Changed Files
+- `src/features/billing/components/CurrentPlanCard.tsx`
+- `src/features/billing/components/PricingCards.tsx`
+
+### Statistics
+```
+ 2 files changed, 13 insertions(+), 6 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-17 17:07:40
 
 ### Changed Files